+++ Rototo/Includes/TweenManager.h
@@ -356,7 +356,7 @@ public:
 	}
 
 	void Init();
-	void Update(Uint64 _elapsed);
+	void Update(unsigned int _elapsed);
 	void Shutdown();
 
 	void AddTween(Tween * _tween);
